pyproject.toml
[project]
name = "karps-backend"
version = "1.0.0-dev"
description = "Lexical infrastructure"
readme = "README.md"
requires-python = ">=3.12.3"
dependencies = [
    "fastapi[standard]>=0.115.6",
    "ruff>=0.9.2",
    "PyYAML==6.0.1",
    "mysql-connector-python",
    "environs",
]
